Retornando a estrutura antiga do axios e removendo as APIs criadas
